+++ conductor-client-spring/src/main/java/com/netflix/conductor/client/spring/ConductorClientAutoConfiguration.java
@@ -85,7 +85,6 @@ public class ConductorClientAutoConfiguration {
                                                      TaskClient taskClient,
                                                      ClientProperties clientProperties,
                                                      List<Worker> workers,
-                                                     Optional<FileClient> fileClient,
                                                      Optional<MetricsCollector> metricsCollector) {
         Map<String, Integer> taskThreadCount = new HashMap<>();
         for (Worker worker : workers) {
@@ -109,7 +108,6 @@ public class ConductorClientAutoConfiguration {
                 .withTaskToDomain(clientProperties.getTaskToDomain())
                 .withShutdownGracePeriodSeconds(clientProperties.getShutdownGracePeriodSeconds())
                 .withTaskPollTimeout(clientProperties.getTaskPollTimeout());
-        fileClient.ifPresent(builder::withFileClient);
         metricsCollector.ifPresent(builder::withMetricsCollector);
         return builder.build();
     }
@@ -140,6 +138,6 @@ public class ConductorClientAutoConfiguration {
     @ConditionalOnBean(ConductorClient.class)
     @ConditionalOnMissingBean
     public FileClient fileClient(ConductorClient client, FileClientProperties fileClientProperties) {
-        return new FileClient(client, fileClientProperties, null);
+        return new FileClient(client, fileClientProperties);
     }
 }